[fix](wal) rename wal_reader and move to wal directory
