+++ be/src/exec/scan/file_scanner.cpp
@@ -78,8 +78,8 @@
 #include "format/table/transactional_hive_reader.h"
 #include "format/table/trino_connector_jni_reader.h"
 #include "format/text/text_reader.h"
-#include "format/wal/wal_reader.h"
 #include "io/cache/block_file_cache_profile.h"
+#include "load/group_commit/wal/wal_reader.h"
 #include "runtime/descriptors.h"
 #include "runtime/runtime_profile.h"
 #include "runtime/runtime_state.h"
