+++ be/src/load/group_commit/wal/wal_table.cpp
@@ -334,7 +334,8 @@ Status WalTable::_get_column_info(int64_t db_id, int64_t tb_id,
 }
 
 Status WalTable::_read_wal_header(const std::string& wal_path, std::string& columns) {
-    std::shared_ptr<doris::WalReader> wal_reader = std::make_shared<WalReader>(wal_path);
+    std::shared_ptr<doris::WalFileReader> wal_reader =
+            std::make_shared<doris::WalFileReader>(wal_path);
     RETURN_IF_ERROR(wal_reader->init());
     uint32_t version = 0;
     RETURN_IF_ERROR(wal_reader->read_header(version, columns));
